Update all permalinks without server-restart
I had an issues in my pages.11tydata.js where the code to rewrite the path to the root only worked on the initial build.

This is updated to use 'path' instead of 'filePathStem' to rewrite the permalink.

At the same time I changed from .html to / to enforce trailing slash on all pages.